Add files via upload
This program will generate 200 random numbers(between 1-1000) and store then in array and then it will ask you to search any number. If your searched number is present in the array it will display the index number and if the number will not there it will appear a sorry message.